src/main/java/br/com/xbrain/autenticacao/modules/usuarioacesso/service/UsuarioAcessoService.java
package br.com.xbrain.autenticacao.modules.usuarioacesso.service;

import br.com.xbrain.autenticacao.modules.usuarioacesso.model.UsuarioAcesso;
import br.com.xbrain.autenticacao.modules.usuarioacesso.repository.UsuarioAcessoRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class UsuarioAcessoService {

    @Autowired
    private UsuarioAcessoRepository usuarioAcessoRepository;

    @Transactional
    public void registrarAcesso(Integer usuarioId) {
        usuarioAcessoRepository.save(UsuarioAcesso.builder()
                .build()
                .criaRegistroAcesso(usuarioId));
    }
}
